models/requirements.txt
fastai
numpy
matplotlib
pillow